requirements.txt
mysql-connector-python~=8.0.25
pandas~=1.2.4
requests~=2.25.1
bs4~=0.0.1
beautifulsoup4~=4.9.3
herepy~=3.5.2
numpy~=1.20.3
geopy~=2.1.0